refactor: moved onboarding request tests to separate file
